core::VBO (+core::IBO) rework, make Buffer template class | Code refactoring

+++ engine/src/over/engine/App.cpp
@@ -33,7 +33,7 @@ void App::Run(int32 width, int32 height, int32 swapInterval) {
   Init();
 
   while (!_window.ShouldClose()) {
-    startTime = glfwGetTime();
+    startTime = static_cast<float32>(glfwGetTime());
 
     _input.Poll();
     _ctx.ClearAll();
